% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{clean_file_path}
\alias{clean_file_path}
\title{Remove artefacts from file path}
\usage{
clean_file_path(file_location)
}
\arguments{
\item{file_location}{Character scalar specifying the path of a file.}
}
\value{
A character scalar
}
\description{
Remove artefacts from file path
}
\examples{
\dontrun{
cleaned_file_path <- clean_file_path(
 file_location = "tests/testthat/example_project/R/calculate_QALYs.R:L41"
)
cleaned_file_path <- clean_file_path(
 file_location = c(
   "tests/testthat/example_project/R/calculate_QALYs.R:L41",
   "tests/testthat/example_project/R/calculate_QALYs.R:L49"
 )
)
}
}
