% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{clean_function_line}
\alias{clean_function_line}
\title{Extract function line in file path}
\usage{
clean_function_line(file_location)
}
\arguments{
\item{file_location}{Character scalar specifying the path of a file.}
}
\value{
A numeric scalar
}
\description{
Extract function line in file path
}
\examples{
\dontrun{
cleaned_function_line <- clean_function_line(
 file_location = "tests/testthat/example_project/R/calculate_QALYs.R:L41"
)
cleaned_function_line <- clean_function_line(
 file_location = c(
   "tests/testthat/example_project/R/calculate_QALYs.R:L41",
   "tests/testthat/example_project/R/calculate_QALYs.R:L49"
 )
)
}
}
