% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{define_app_ui}
\alias{define_app_ui}
\title{Create Shiny app UI}
\usage{
define_app_ui(network_title)
}
\arguments{
\item{network_title}{Character string representing the title of the network to be displayed above the network.}
}
\value{
Shiny app user interface
}
\description{
Create Shiny app UI
}
