% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{filter_func_lnums}
\alias{filter_func_lnums}
\title{Helper function for find_function_definitions}
\usage{
filter_func_lnums(v_func_lnums, v_assign_lnums, v_symbol_lnums, parsed_data)
}
\description{
Do not try to call this - it is only a help
and is only to keep find_function_definitions clean
These two functions could really be combined - a task for another day
}
