% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{find_function_definitions}
\alias{find_function_definitions}
\title{Parses an R source file, returns function names defined within.}
\usage{
find_function_definitions(filename)
}
\arguments{
\item{filename}{A string containing a path to an R source file}
}
\value{
A dataframe with interesting information
}
\description{
Using utils::getParseData(), searches for function definitions
by matching the FUNCTION keyword (i.e. "function")
with it's associated SYMBOL (i.e the function name)
}
\examples{
file_path <- assertHE_example("example_scripts/example_tricky_functions.R")
find_function_definitions(filename = file_path)

}
