% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{find_next_vector_element}
\alias{find_next_vector_element}
\title{Find the next element of the vector after a value}
\usage{
find_next_vector_element(value, vector, LTE = FALSE)
}
\arguments{
\item{value}{A value of numeric values}

\item{vector}{A vector of numeric values}

\item{LTE}{a boolean to determine collection on "greater than or equal"}
}
\value{
The next element of the vector after the value
}
\description{
Find the next element of the vector after a value
}
\examples{
find_next_vector_element(value = 5, vector = 1:10)
find_next_vector_element(value = 5, vector = 1:4)
find_next_vector_element(value = 5, vector = 1:5, LTE = FALSE)
find_next_vector_element(value = 5, vector = 1:5, LTE = TRUE)

}
