% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{find_test}
\alias{find_test}
\title{Find test for a function in a codebase}
\usage{
find_test(v_functions, path = ".", test_path = "tests/testthat")
}
\arguments{
\item{v_functions}{A vector of functions to search for.}

\item{path}{The path to the folder to be checked.}

\item{test_path}{The relative path to the test folder from the project folder (path).}
}
\value{
A vector of file paths to the test scripts for each function.
Returns NA where no script can be found.
}
\description{
This function finds the test for each in a vector of functions in a specified
testing folder, default = tests/testthat as the relative path from the project folder (path).
}
\examples{
\dontrun{
 v_funcs_to_find_tests_for <- c("check_init", "mean", "check_markov_trace", "find_test")
 find_test(v_functions = v_funcs_to_find_tests_for,
           path = ".")
}

}
\seealso{
Other checking_functions: 
\code{\link{list_functions_in_script}()},
\code{\link{tabulate_functions_in_folder}()},
\code{\link{tabulate_functions_in_script}()}
}
\concept{checking_functions}
