% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{get_active_functions}
\alias{get_active_functions}
\title{get all active functions that exist in the global environment}
\usage{
get_active_functions(packages = "assertHE")
}
\arguments{
\item{packages}{a vector containing the names of packages to include in the search}
}
\value{
a vector containing the names of all active functions in the global environment
}
\description{
get all active functions that exist in the global environment
}
