% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_foo_coverage.R
\name{get_foo_coverage}
\alias{get_foo_coverage}
\title{Get coverage by function}
\usage{
get_foo_coverage(foo_folder, test_folder)
}
\arguments{
\item{foo_folder}{folder containing functions}

\item{test_folder}{folder containing tests}
}
\value{
a dataframe with a column for functions and a column for coverage
}
\description{
Get coverage by function
}
\examples{
\donttest{
# Example takes more than 5 seconds to run
if(require(testthat)) {
  folder_path1 <- assertHE_example("example_project/R")
  folder_path2 <- assertHE_example("example_project/tests/testthat")
  get_foo_coverage(
    foo_folder = folder_path1,
    test_folder = folder_path2
  )
}
}

}
