% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{get_isolated_foo}
\alias{get_isolated_foo}
\title{Get Isolated Functions}
\usage{
get_isolated_foo(df_edges)
}
\arguments{
\item{df_edges}{A dataframe with two columns ("from" and "to") representing the dependencies.}
}
\value{
A vector of isolated function names.
}
\description{
Get Isolated Functions
}
