% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_function_with_LLM.R
\name{get_roxygen_description}
\alias{get_roxygen_description}
\title{Get Title and Description from Parsed List}
\usage{
get_roxygen_description(parsed_list)
}
\arguments{
\item{parsed_list}{A list containing parsed elements.}
}
\value{
A list containing the title and description.
}
\description{
This function extracts the title and description from a parsed list.
}
\examples{
parsed_list <- list(list(tag = "title", val = "Sample Title"),
                     list(tag = "description", val = "This is a sample description."))
get_roxygen_description(parsed_list)

}
