% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_function_with_LLM.R
\name{get_roxygen_description_from_foo}
\alias{get_roxygen_description_from_foo}
\title{Get roxygen title and description from function}
\usage{
get_roxygen_description_from_foo(foo_name)
}
\arguments{
\item{foo_name}{function for which want descriotion}
}
\value{
text containing description
}
\description{
Get roxygen title and description from function
}
