% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Summarise project functions with details on packages and existence of unit-tests.}
\usage{
get_summary_table(
  path = ".",
  path_exclude = "testthat/",
  test_path = "tests/",
  cheers_pattern = "@family",
  packages_to_exclude = c("base", "stats", "utils")
)
}
\arguments{
\item{path}{The path to the folder to be checked.}

\item{path_exclude}{A set of strings that will exclude any files if it is present in the file path.}

\item{test_path}{The relative path to the test folder from the project folder (path).}

\item{cheers_pattern}{A string that will be used to identify the cheers tag.}

\item{packages_to_exclude}{A vector of packages to exclude from the search.}
}
\value{
A dataframe with the following columns:
\itemize{
\item function_name: The name of the function.
\item package: The package the function is from.
\item file_name: The file in which the function is defined.
\item test_location: The file in which the function is tested.
}
}
\description{
Creates a summary table containing the name of each function in the project,
the package it is from, whether it has a unit-test and the file in which it is tested.
}
\examples{
\dontrun{
get_summary_table()
}

}
