% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\docType{data}
\name{help_text}
\alias{help_text}
\title{Help text}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
help_text
}
\value{
A shiny HTML element
}
\description{
Help text
}
\keyword{datasets}
