% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_undefined_function_objects.R
\name{identify_undefined_function_objects}
\alias{identify_undefined_function_objects}
\title{Function to check for undefined objects in the function body}
\usage{
identify_undefined_function_objects(func, env = .GlobalEnv)
}
\arguments{
\item{func}{A function object}

\item{env}{The environment in which to evaluate the function. Default is '.GlobalEnv'}
}
\value{
A character vector of undefined objects contained in the function body
}
\description{
Function to check for undefined objects in the function body
}
\examples{

# Example function
noArgFunction <- function(y) {
 x <- undefined_object1 * y
 return(x)
 }

noArgFunction2 <- function(y) {
 undefined_object2 <- undefined_object2
 x <- undefined_object2 * y
 return(x)
 }

 # Check for undefined objects
 identify_undefined_function_objects("noArgFunction")

 # currently fails for the second example!!!
 identify_undefined_function_objects(noArgFunction2)

}
