% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_undefined_function_objects.R
\name{identify_undefined_function_objects_in_folder}
\alias{identify_undefined_function_objects_in_folder}
\title{Function to source all files in folder & check for undefined objects}
\usage{
identify_undefined_function_objects_in_folder(foo_folder)
}
\arguments{
\item{foo_folder}{A character vector of the folder path}
}
\value{
A list of character vectors of undefined objects in each function in a folder
}
\description{
It loads the functions to the global environment, which is not ideal.
But it works for now.
}
\examples{
\dontrun{

foo_folder  <- testthat::test_path("example_project/R")
test_folder <- testthat::test_path("example_project/tests/testthat")

identify_undefined_function_objects_in_folder(foo_folder)
}
}
