% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{list_functions_in_script}
\alias{list_functions_in_script}
\title{List all of the functions in a script}
\usage{
list_functions_in_script(filename, alphabetic = TRUE)
}
\arguments{
\item{filename}{The name of the file to be checked.}

\item{alphabetic}{If TRUE, return the functions in alphabetic order.}
}
\value{
A named list of functions in the script, by package.
}
\description{
This function lists all of the functions in a script. It expects the script to be
an R file, and requires that package libraries are loaded when identifying packages.
}
