% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{processNodes}
\alias{processNodes}
\title{Process Nodes}
\usage{
processNodes(df_edges, from_col = "from", to_col = "to")
}
\arguments{
\item{df_edges}{A data frame containing columns "from" and "to" representing the edges of the network.}

\item{from_col}{Name of the column in df_edges representing the source nodes.}

\item{to_col}{Name of the column in df_edges representing the target nodes.}
}
\value{
A data frame of unique nodes with labels.
}
\description{
Process unique nodes from a dataframe of edges.
}
