% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_files}
\alias{source_files}
\title{Source all files in target folder}
\usage{
source_files(
  path = ".",
  dirExclRegx = NULL,
  fileIncRegx = "\\\\.R$",
  fileExclRegx = NULL,
  recursive = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory. Defaults to current working directory.}

\item{dirExclRegx}{Character string containing regular expression for directories to exclude.}

\item{fileIncRegx}{Character string containing regular expression for files to include. Defaults to '\\.R$' for R files.}

\item{fileExclRegx}{Character string containing regular expression for files to exclude.}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{verbose}{logical. Whether to print files as they are being sourced.}
}
\value{
nothing, sources functions to global environment.
}
\description{
Recursively source all files in the target directory.
}
