% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{source_funcs}
\alias{source_funcs}
\title{source_funcs}
\usage{
source_funcs(file, env)
}
\arguments{
\item{file}{a connection object or a character string path to a file.}

\item{env}{the environment in which to source the functions.}
}
\value{
No return value, called for side effects.
}
\description{
Sources \emph{only} the functions discovered in an R file.
}
\section{IMPORTANT !!!}{
Sourcing \emph{this} file is a mistake - may result in infinite recursion.
}

