% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{tabulate_functions_in_folder}
\alias{tabulate_functions_in_folder}
\title{Create table of all of the functions identified in a project folder}
\usage{
tabulate_functions_in_folder(
  path = ".",
  collapse = T,
  packages_to_exclude = c("base", "stats", "utils")
)
}
\arguments{
\item{path}{The path to the folder to be checked.}

\item{collapse}{If TRUE, return a single data-frame of all functions. Else return a list by file.}

\item{packages_to_exclude}{A vector of packages to exclude from the output (e.g. "base")}
}
\value{
Either a data-frame of functions and the package the function is from, or a list of functions by file.
}
\description{
This function tabulates all of the functions identified in R scripts within a
project folder. It requires that package libraries are loaded when identifying
a function's package.
}
\examples{

\dontrun{
tabulate_functions_in_folder(
    path = ".",
    collapse = T,
    packages_to_exclude = c("base", "stats", "utils")
    )
}
}
