% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{tabulate_functions_in_folder_with_tests}
\alias{tabulate_functions_in_folder_with_tests}
\title{Summarise project functions with details on packages and existence of unit-tests.}
\usage{
tabulate_functions_in_folder_with_tests(
  path = ".",
  path_exclude = "testthat/",
  packages_to_exclude = .packages(TRUE),
  test_path = "tests/testthat"
)
}
\arguments{
\item{path}{The path to the folder to be checked.}

\item{path_exclude}{A set of strings that will exclude any files if it is present in the file path.}

\item{packages_to_exclude}{A vector of packages to exclude from the search.}

\item{test_path}{The relative path to the test folder from the project folder (path).}
}
\value{
A dataframe with the following columns:
\itemize{
\item function_name: The name of the function.
\item package: The package the function is from.
\item file_name: The file in which the function is defined.
\item test_location: The file in which the function is tested.
}
}
\description{
Creates a summary table containing the name of each function in the project,
the package it is from, whether it has a unit-test and the file in which it is tested.
}
\examples{
\dontrun{
tabulate_functions_in_folder_with_tests(
path = ".",
path_exclude = "testthat/",
packages_to_exclude = c("base", "stats", "ggplot2"),
test_path = "tests/testthat"
)
}

}
