% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{tabulate_functions_in_script}
\alias{tabulate_functions_in_script}
\title{Create table of all of the functions in a script}
\usage{
tabulate_functions_in_script(
  filename,
  packages_to_exclude = c("base", "stats", "utils")
)
}
\arguments{
\item{filename}{The name of the file to be checked.}

\item{packages_to_exclude}{A vector of packages to exclude from the output (e.g. "base")}
}
\value{
A data-frame of functions and the package the function is from.
}
\description{
This function tabulates all of the functions in a script. It expects the script to be
an R file, and requires that package libraries are loaded when identifying packages.
}
