% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{assertHE_example}
\alias{assertHE_example}
\title{Get path to assertHE example}
\usage{
assertHE_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
assertHE comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
assertHE_example()
assertHE_example("challenge.csv")
}
