% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_function_with_LLM.R
\name{create_prompt}
\alias{create_prompt}
\title{Create a prompt for a LLM}
\usage{
create_prompt(
  foo_arguments,
  foo_body,
  foo_name,
  foo_desc,
  foo_title,
  text_language = "English"
)
}
\arguments{
\item{foo_arguments}{the arguments to the function}

\item{foo_body}{the body of the function}

\item{foo_name}{function name}

\item{foo_desc}{function description}

\item{foo_title}{function title}

\item{text_language}{the language to return the summary in}
}
\value{
a single prompt in the form of a character string
}
\description{
Uses the function arguments and function body as inputs to create a prompt
for the LLM.
}
\examples{
\dontrun{
create_prompt(
text_language = "Spanish",
foo_arguments = LETTERS[1:3],
foo_body = "D <- A+B+C; return(D)",
foo_name = "calculate_QALYs",
foo_desc = "This function calcs QALYs",
foo_title = "Calculate the QALYs")
}
}
