% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{define_app_server}
\alias{define_app_server}
\title{Create Shiny app server logic}
\usage{
define_app_server(network_object, project_path, foo_path)
}
\arguments{
\item{network_object}{visNetwork object to be displayed in the shiny app}

\item{project_path}{Path to the project directory}

\item{foo_path}{path to the function folder}
}
\value{
Shiny app server logic
}
\description{
Create Shiny app server logic
}
