% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{define_app_server}
\alias{define_app_server}
\title{Create Shiny app server logic}
\usage{
define_app_server(network_object)
}
\arguments{
\item{network_object}{visNetwork object to be displayed in the shiny app}
}
\value{
Shiny app server logic
}
\description{
Create Shiny app server logic
}
