% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{extract_function_name}
\alias{extract_function_name}
\title{Extract function name from a string}
\usage{
extract_function_name(string)
}
\arguments{
\item{string}{A string containing a function definition, this must contain the word 'function'}
}
\value{
A string containing the function name
}
\description{
Extract function name from a long string. This works by
splitting the string by "<-", "=", and " " and then keeping the word
before the first of these characters.
}
