% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{find_folder_function_definitions}
\alias{find_folder_function_definitions}
\title{Creates summary of R files in folder with functions defined within and locations.}
\usage{
find_folder_function_definitions(foo_folder)
}
\arguments{
\item{foo_folder}{A folder to apply find_function_definitions to each script in.}
}
\value{
A dataframe containing a column for function string and a column for function location.
}
\description{
Applies find_function_definitions to each file in a folder and aggregate results
}
\examples{
\dontrun{
find_folder_function_definitions(
foo_folder = "tests/testthat/example_project"
)
}
}
