% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{find_next_vector_element}
\alias{find_next_vector_element}
\title{Find the next element of the vector after a value}
\usage{
find_next_vector_element(value, vector)
}
\arguments{
\item{value}{A value of numeric values}

\item{vector}{A vector of numeric values}
}
\value{
The next element of the vector after the value
}
\description{
Find the next element of the vector after a value
}
\examples{
\dontrun{
find_next_vector_element(value = 5, vector = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}

}
