% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{find_previous_vector_element}
\alias{find_previous_vector_element}
\title{Find the previous element of the vector before a value}
\usage{
find_previous_vector_element(value, vector, LTE = FALSE)
}
\arguments{
\item{value}{A value of numeric values}

\item{vector}{A vector of numeric values}

\item{LTE}{a boolean to determine collection on "less than" or "less than equal"}
}
\value{
The previous element of the vector before the value
}
\description{
Find the previous element of the vector before a value
}
\examples{
find_previous_vector_element(value = 5, vector = 1:10)
find_previous_vector_element(value = 5, vector = 6:10)
find_previous_vector_element(value = 5, vector = 5:10, LTE = FALSE)
find_previous_vector_element(value = 5, vector = 5:10, LTE = TRUE)

}
