% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{get_file_cheers_classifications}
\alias{get_file_cheers_classifications}
\title{Get cheers classification tags from a given file}
\usage{
get_file_cheers_classifications(
  filename,
  cheers_pattern,
  function_pattern = "function\\\\("
)
}
\arguments{
\item{filename}{A string containing the filepath to the file to be checked}

\item{cheers_pattern}{A string containing the roxygen tag for cheers which is used as an identifier}

\item{function_pattern}{A string containing the pattern to identify functions}
}
\value{
A list containing the cheers tags and the function names that follow them
}
\description{
For a provided filepath, identify the cheers classification tags
and the function names that follow them.
}
\seealso{
Other cheers: 
\code{\link{get_folder_cheers_classifications}()}
}
\concept{cheers}
