% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheers_checker.R
\name{get_folder_cheers_classifications}
\alias{get_folder_cheers_classifications}
\title{Get cheers classification tags from a given folder}
\usage{
get_folder_cheers_classifications(path, cheers_pattern, path_ignore = "tests/")
}
\arguments{
\item{path}{A string containing the filepath to the folder to be checked}

\item{cheers_pattern}{A string containing the roxygen tag for cheers which is used as an identifier}

\item{path_ignore}{A string containing the pattern to identify files to ignore}
}
\value{
A list containing the cheers tags and the function names that follow them
}
\description{
For a provided folder path, identify the cheers classification tags
and the function names that follow them.
}
\seealso{
Other cheers: 
\code{\link{get_file_cheers_classifications}()}
}
\concept{cheers}
