% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{get_function_line}
\alias{get_function_line}
\title{Extract function line in file path}
\usage{
get_function_line(file_location)
}
\arguments{
\item{file_location}{Character scalar specifying the path of a file.}
}
\value{
A numeric scalar
}
\description{
Extract function line in file path
}
