% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{get_function_path}
\alias{get_function_path}
\title{Remove artefacts from file path}
\usage{
get_function_path(file_location, project_path)
}
\arguments{
\item{file_location}{Character scalar specifying the path of a file.}

\item{project_path}{Character scalar specifying the path of the project.}
}
\value{
A character scalar
}
\description{
Remove artefacts from file path
}
