% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{identify_dependencies}
\alias{identify_dependencies}
\title{Identify Dependencies}
\usage{
identify_dependencies(v_unique_foo, pkg_env = environment())
}
\arguments{
\item{v_unique_foo}{Vector of unique function strings.}

\item{pkg_env}{The package environment where the functions are defined
(e.g. global).}
}
\value{
A dataframe with two columns ("from" and "to") representing the dependencies.
}
\description{
Identify dependencies between functions.
}
