% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{locate_funcs}
\alias{locate_funcs}
\title{locate_funcs}
\usage{
locate_funcs(file)
}
\arguments{
\item{file}{= a connection object or a character string path to a file.}
}
\value{
Returns a data frame with the following columns:
func_num: The ID of the function - monotonic increasing from 1.
func_start: The line number (within the file) of the function start.
func_end: The line number of the function end.
}
\description{
locates the lines which define a function within a single file
}
\examples{
\dontrun{
locate_funcs(file)
}

}
