% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot Network}
\usage{
plotNetwork(
  df_edges,
  from_col = "from",
  to_col = "to",
  df_summary,
  df_coverage
)
}
\arguments{
\item{df_edges}{A data frame containing columns "from" and "to" representing the edges of the network.}

\item{from_col}{Name of the column in df_edges representing the source nodes.}

\item{to_col}{Name of the column in df_edges representing the target nodes.}

\item{df_summary}{A summary dataframe containing the information about each function.}

\item{df_coverage}{a summary dataframe with function names and test coverages}
}
\value{
A visNetwork object representing the network plot.
}
\description{
Visualize a network plot using the visNetwork package.
}
\examples{
\dontrun{
# Plot a network from a data frame of edges
plotNetwork(df_edges)
}

}
