% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot Network}
\usage{
plotNetwork(df_edges, from_col = "from", to_col = "to")
}
\arguments{
\item{df_edges}{A data frame containing columns "from" and "to" representing the edges of the network.}

\item{from_col}{Name of the column in df_edges representing the source nodes.}

\item{to_col}{Name of the column in df_edges representing the target nodes.}
}
\value{
A visNetwork object representing the network plot.
}
\description{
Visualize a network plot using the visNetwork package.
}
\examples{
\dontrun{
# Plot a network from a data frame of edges
plotNetwork(df_edges)
}

}
