% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_function_with_LLM.R
\name{return_message}
\alias{return_message}
\title{Extract the content from the output of the LLM}
\usage{
return_message(API_response, verbose = T)
}
\arguments{
\item{API_response}{response from the LLM API}

\item{verbose}{whether to include the message for the number of token's used}
}
\value{
A single string summary of the content of the LLM response
}
\description{
Extracts content and prints the number of tokens used as a message.
}
