% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_visualiser.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run a Shiny app to host a network visualization}
\usage{
run_shiny_app(
  uiFunction = define_app_ui,
  serverFunction = define_app_server,
  network_object
)
}
\arguments{
\item{uiFunction}{Function defining shiny user-interface}

\item{serverFunction}{Function defining shiny server logic}

\item{network_object}{visNetwork object to be displayed in the shiny app}
}
\value{
A shiny app
}
\description{
Run a Shiny app to host a network visualization
}
\examples{
\dontrun {
network_object <- visualise_project(
    project_path = "tests/testthat/example_project",
    foo_path = "R",
    test_path = "tests/testthat",
    run_coverage = TRUE
 )

 run_shiny_app(network_object = network_object)
}
}
