% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{source_files}
\alias{source_files}
\title{source_files}
\usage{
source_files(
  file_regx = ".R",
  path = ".",
  recursive = TRUE,
  exclude_files = NULL,
  exclude_dirs = NULL,
  funcs_only = FALSE,
  verbose = FALSE,
  keep_source = FALSE
)
}
\arguments{
\item{file_regx}{= ".*" - a regular expression for files to source}

\item{path}{= "." - a path to search}

\item{recursive}{= TRUE - recurse into subdirectories}

\item{exclude_files}{= NULL - regx for files to exclude}

\item{exclude_dirs}{= NULL - regx for directories to exclude}

\item{funcs_only}{= FALSE - source \emph{only} identified functions, not the whole file}

\item{verbose}{= FALSE - whether to emit the sourced files.}

\item{keep_source}{= FALSE - whether to keep the source data when using source.}
}
\value{
list of files sourced
}
\description{
Source files based upon regular expression searching
}
\section{IMPORTANT !!!}{
sourcing \emph{this} file is a mistake - may result in infinite recursion
}

\examples{
\dontrun{
source_files(file_regx = ".*",  ## any file name
 path = ".*",   # the current directory and all subdirectories
 recursive = FALSE,  # don't recurse
 exclude_files = ".*utility.*", # exclude "utility" anywhere in basename
 exclude_dirs = "\\\\<tmp\\\\>|/tmp/|/tmp\\\\>|\\\\<tmp/"  # exclude any directory named "tmp", or subdirs
 )
}

}
