% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file.R
\name{source_lines}
\alias{source_lines}
\title{source_lines}
\usage{
source_lines(file, lines, env)
}
\arguments{
\item{file}{a connection object or a character string path to a file.}

\item{lines}{A vector of integers specifying the lines to be sourced.}

\item{env}{the environment in which to source the lines.}
}
\value{
No return value, called for side effects.
}
\description{
Sources specified lines within a single file.
}
\section{IMPORTANT !!!}{
Sourcing \emph{this} file is a mistake - may result in infinite recursion.
}

\examples{
\dontrun{
source_lines(file = "PATH/filename",  ## any file name, or a Connection object
             lines = c(4, 5, 6) )     ## source lines 4-6
}

}
