% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_function_with_LLM.R
\name{summarise_function_with_LLM}
\alias{summarise_function_with_LLM}
\title{Summarize a function using a Large Language Model}
\usage{
summarise_function_with_LLM(foo_name)
}
\arguments{
\item{foo_name}{function name}
}
\value{
A character string with a summary of the function based on its arguments and body.
}
\description{
This function summarizes another function using a Language Model.
}
\examples{
\dontrun{
summarise_function_with_LLM("calculate_costs")
}

}
