% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_finder.R
\name{summarise_model}
\alias{summarise_model}
\title{Summarise the model functions in a single folder.}
\usage{
summarise_model(foo_folder, test_folder = NULL, output_format = "dataframe")
}
\arguments{
\item{foo_folder}{path to folder containing all functions for the model}

\item{test_folder}{folder containing all tests}

\item{output_format}{output format to use, defaults to dataframe, options include latex and word.}
}
\value{
dataframe with three columns. 'foo_string' contains function names, 'foo_location'
contains the location of the function definitions, 'test_location' contains the locations
of tests for each function (both file and line number).
}
\description{
Summarise the model functions in a single folder.
}
\examples{
\dontrun{
foo_folder  <- "C:/Users/r_a_s/Documents/Projects/GSK/assertHE/tests/testthat/example_project/R"
test_folder <- "./tests/testthat/example_project/tests/testthat"

summarise_model(foo_folder = foo_folder, test_folder =  test_folder)

summarise_model(foo_folder = foo_folder, test_folder =  NULL)

}
}
